--- extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/crawler/UpdateResponse.java
@@ -1,4 +1,0 @@
-package org.eclipse.dataspaceconnector.crawler;
-
-public class UpdateResponse {
-}
