+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/crawler/spi/ProtocolAdapter.java
@@ -1,7 +1,7 @@
-package org.eclipse.dataspaceconnector.crawler;
+package org.eclipse.dataspaceconnector.crawler.spi;
 
 import java.util.concurrent.CompletableFuture;
 
-public interface CrawlerProtocolAdapter {
+public interface ProtocolAdapter {
     CompletableFuture<UpdateResponse> sendRequest(UpdateRequest request);
 }
