+++ extensions/catalog/federated-catalog-cache/src/main/java/org/eclipse/dataspaceconnector/crawler/spi/UpdateResponse.java
@@ -1,0 +1,4 @@
+package org.eclipse.dataspaceconnector.crawler.spi;
+
+public class UpdateResponse {
+}
